WebAuthn: Fork in-progress upstream provider
This will give us a starting point for our version of the provider. The files were copied verbatim, and then minor changes were made to adjust file paths etc and get it working in this plugin. Future commits will swap in the MadWizard library, and make other changes to fit w.org needs.

Co-Authored-By: mcguffin <joern@flyingletters.com>