Return `save()` promise so callers can wait for result
Some UI interactions depend on the save completing before they should be done. For example, when registering a new WebAuthn key, we shouldn't redirect the user to their list of keys until the new key has been saved to the database. Otherwise it wouldn't show up in the list at first, and then would appear later.

Co-Authored-By: Ren <18050944+renintw@users.noreply.github.com>