File structure and components refactoring and test update (#215)
* Rename the original utilities.js to common.js to indicate its generic use

* Move the customized hook "useUser" to hooks/

* Move tests/ directory outside of components/

* Separate useUser from common.test

* Add renderHook for useUser test.

The hook should be called by a component. In a test, renderHook would do that for us.

* Extract ScreenNavigation

* Add navigateToScreen as dependency

navigateToScreen should be added as the hook dependency, or hasEdits might not be the latest value when calling the function. In fact, due to the property of how hasEdits is updated, the navigateToScreen would always get the latest value of hasEdits, but it's best practice to add state and props that are used in the hook as its dependency, or it might get harder to develop a new feature or maintain the code someday in the future.

* Bump @wordpress/scripts version to ^25.0.0

See details at https://github.com/WordPress/wporg-two-factor/pull/215#issue-1748345272